requirements.txt
python-bitcoinrpc
python-dotenv
pytest #optional